extensions refactored, add new driver for tests
