+++ test/GdalCore-XUnit/GdalCore-XUnit.csproj
@@ -1,30 +1,27 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
+  <PropertyGroup> 
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <RootNamespace>GdalCore_XUnit</RootNamespace>
+    <Platform>x64</Platform>
     <NoWarn>NU1605</NoWarn>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
-
   <ItemGroup>
-    <PackageReference Include="MaxRev.Gdal.Core" Version="3.2.0.250" />
-    <PackageReference Include="MaxRev.Gdal.LinuxRuntime.Minimal" Version="3.2.0.250" />
-    <PackageReference Include="MaxRev.Gdal.WindowsRuntime.Minimal" Version="3.2.0.250" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
+    <PackageReference Include="MaxRev.Gdal.Core" Version="3.2.0.300" />
+    <PackageReference Include="MaxRev.Gdal.LinuxRuntime.Minimal" Version="3.2.0.300" />
+    <PackageReference Include="MaxRev.Gdal.WindowsRuntime.Minimal" Version="3.2.0.300" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
     <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.2">
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="coverlet.collector" Version="1.3.0">
+    <PackageReference Include="coverlet.collector" Version="3.0.2">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
   </ItemGroup>
-
   <ItemGroup>
     <ProjectReference Include="..\GdalCoreTest\GdalCoreTest.csproj" />
   </ItemGroup>
-
 </Project>
